% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_tables_md.R
\name{format_tables_md}
\alias{format_tables_md}
\title{Format Markdown Tables}
\usage{
format_tables_md(df, caption = NA)
}
\arguments{
\item{df}{A data frame to view as a table}

\item{caption}{A caption, if needed}
}
\value{
A formatted data frame to display in a markdown document
}
\description{
Format Markdown Tables
}
\examples{
\dontrun{format_tables(mtcars)}
}
