% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_ecdf.R
\name{max_ecdf}
\alias{max_ecdf}
\title{Cumulative Distribution Max Difference Value Identifier}
\usage{
max_ecdf(df, group_var, feature_var)
}
\arguments{
\item{df}{A data frame}

\item{group_var}{The variable that contains your two groups that you want to
compare--think 1 vs. 0}

\item{feature_var}{The feature or variable whose values you are interested
in exploring}
}
\value{
A tibble with the feature_var value where the cumulative
diffference is largest. Includes the positive and negative values as well.
}
\description{
Cumulative Distribution Max Difference Value Identifier
}
\examples{
\dontrun{max_ecdf(df, outcome, key)}
}
