% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_h2o_models_metrics.R
\name{save_h2o_models_metrics}
\alias{save_h2o_models_metrics}
\title{Save models and various performance metrics from h2o autoML run}
\usage{
save_h2o_models_metrics(leaderboard, max_top_models = NA, path_slug)
}
\arguments{
\item{leaderboard}{an h2o leaderboard object}

\item{max_top_models}{The top-n models to save. Defaults to NA, which
saves all}

\item{path_slug}{A path where you want the models to be saved}
}
\value{

}
\description{
Save models and various performance metrics from h2o autoML run
}
\examples{
\dontrun{save_h2o_models_metrics(leaderbard_a, max_top_models = 5, path_slug = getwd())}
}
