% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{Send email from within R}
\usage{
send_email(to_email_address, cc_email_address = NULL,
  bcc_email_address = NULL, from_email_address = NULL,
  from_email_password, from_email_username = NULL,
  attachment_path = NULL, email_subject = "", email_body = "",
  email_body_type = "plain", from_host = "smtp.gmail.com",
  port = 465)
}
\arguments{
\item{to_email_address}{A recipient email address.}

\item{cc_email_address}{A recipient to copy.}

\item{bcc_email_address}{A recipient to blind copy.}

\item{from_email_address}{Email address to send from.}

\item{from_email_password}{Password for the from_email_address.}

\item{from_email_username}{Username for the from_email_address. If no argument provided the
from_email_address will be used.}

\item{attachment_path}{A path to a file to attach to the outgoing message.}

\item{email_subject}{A subject for the outgoing message.}

\item{email_body}{The body of the outgoing message.}

\item{email_body_type}{Either "plain" or "html". Pass "html" to use custom html formatting
in the body of the message}

\item{from_host}{from_email_address host}

\item{port}{Port to send the outgoing message through}
}
\value{
Sends an email with content based on the arguments provided
}
\description{
Send email from within R
}
\examples{
\dontrun{send_email(email_body = 'Hello world')}
}
