% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_list.R
\name{variable_list}
\alias{variable_list}
\title{Variable List}
\usage{
variable_list(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Gives variable number, name, first observation, and the variable's class and returns a dataframe
}
\description{
Variable List
}
\examples{
\dontrun{variable_list(mtcars)}
}
